rename apply_to_collection to apply_func
